Trying to reproduce jcjohnson/torch-rnn.
Trying to reproduce default setting of jcjohnson/torch-rnn
as closely as possible. Can't be done exactly because we don't
have support to mini-batches. LSTM topology is probably not the same.
Not sure if loss calculation and Adam setting are equivalent. I don't
know either if sequences are shuffled there like I am doing here.